+++ pkgs/shelf_static/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 1.1.3-wip
 
+* `Response`:
+  * Populate `context` with the file system paths used for resolution.
+    * Possible entries: `shelf_static:file`, `shelf_static:file_not_found`, and `shelf_static:directory`.
+
 * Require Dart `^3.3.0`.
 
 ## 1.1.2